+++ routeros/resource_capsman_access_list.go
@@ -79,6 +79,7 @@ func ResourceCapsManAccessList() *schema.Resource {
 			Type:        schema.TypeString,
 			Optional:    true,
 			Description: "PSK passphrase for the client if some PSK authentication algorithm is used.",
+			Sensitive:   true,
 		},
 		"radius_accounting": {
 			Type:        schema.TypeBool,