+++ routeros/resource_system_ntp_server.go
@@ -31,6 +31,7 @@ func ResourceSystemNtpServer() *schema.Resource {
 			Description: "NTP symmetric key, used for authentication between the NTP client and server. Key Identifier " +
 				"(Key ID) - an integer identifying the cryptographic key used to generate the message-authentication code.",
 			DiffSuppressFunc: AlwaysPresentNotUserProvided,
+			Sensitive:        true,
 		},
 		"broadcast": {
 			Type:             schema.TypeBool,